# Utf8JsonReader should store if parsed number contains exponent