--- a/src/System.Text.Json/src/System/Text/Json/JsonReaderState.cs
+++ b/src/System.Text.Json/src/System/Text/Json/JsonReaderState.cs
@@ -22,6 +22,7 @@ namespace System.Text.Json
         internal int _maxDepth;
         internal bool _inObject;
         internal bool _isNotPrimitive;
+        internal bool _numberHasExponent;
         internal JsonTokenType _tokenType;
         internal JsonTokenType _previousTokenType;
         internal JsonReaderOptions _readerOptions;
@@ -69,6 +70,7 @@ namespace System.Text.Json
             _maxDepth = maxDepth;
             _inObject = default;
             _isNotPrimitive = default;
+            _numberHasExponent = default;
             _tokenType = default;
             _previousTokenType = default;
             _readerOptions = options;