--- a/src/System.Text.Json/src/System/Text/Json/Utf8JsonReader.MultiSegment.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Utf8JsonReader.MultiSegment.cs
@@ -34,6 +34,7 @@ namespace System.Text.Json
             _maxDepth = state._maxDepth == 0 ? JsonReaderState.DefaultMaxDepth : state._maxDepth; // If max depth is not set, revert to the default depth.
             _inObject = state._inObject;
             _isNotPrimitive = state._isNotPrimitive;
+            _numberHasExponent = state._numberHasExponent;
             _tokenType = state._tokenType;
             _previousTokenType = state._previousTokenType;
             _readerOptions = state._readerOptions;
@@ -1108,6 +1109,7 @@ namespace System.Text.Json
             // TODO: https://github.com/dotnet/corefx/issues/33294
             Debug.Assert(data.Length > 0);
 
+            _numberHasExponent = false;
             SequencePosition startPosition = _currentPosition;
             int startConsumed = _consumed;
             consumed = 0;
@@ -1199,6 +1201,7 @@ namespace System.Text.Json
 
             Debug.Assert(nextByte == 'E' || nextByte == 'e');
             i++;
+            _numberHasExponent = true;
             _bytePositionInLine++;
 
             signResult = ConsumeSignMultiSegment(ref data, ref i);