Partially addresses: https://github.com/dotnet/corefx/issues/33294#issuecomment-447984832

Stores, whether the most recently parsed number contains an exponent. This avoids another pass over the span to determine the format for the Utf8Parser.

/cc @ahsonkhan