@tkp1n 2019-01-07
@ahsonkhan I just managed to write and run the benchmarks you asked for. I could not see a difference in parsing `{"hello": "world"}` between master and PR outside of error margins. Parsing an array of 1000 10-digit numbers with decimal separators in random locations yielded the following results:

Method |     Mean |     Error |    StdDev |   Median |      Min |      Max |
------------ |---------:|----------:|----------:|---------:|---------:|---------:|
 master      | 194.4 us |  1.690 us |  1.581 us | 194.0 us | 191.7 us | 196.9 us |
 PR          | 182.5 us | 3.6240 us | 4.0281 us | 182.1 us | 177.2 us | 190.9 us |

It looks like a 6% perf boost with no measurable impact due to the larger struct size. 
You could of course always use some bit twiddling to save some bytes on the struct fields, but it does not seem necessary...